Fix bug when old config is restored
There was still some old code left in the createNewInstance function,
that would cause troubles. If the createNewInstance now sees an existing
instance it will trust the reconcile function to make it work.

The reconcileSingleShinyProxyInstance function now also refreshes the SP
instance to make sure that the latest flag of an instance is always up
to date. Otherwise, the instance would not be found.